refactor(storage): keep the reconstruction constructor in the writer module
Ref polylogue-1fijp

insert_reconstructed_raw_row issues an INSERT, and raw_admission.py's own
contract is that it composes source_write's writers rather than
re-implementing raw-row persistence -- devtools verify layering caught the
contradiction as writer_module_unmarked_mutation. Move the constructor to
source_write.py, this tier's declared writer module, and leave a pointer
to it in raw_admission's docstring so the exemption stays discoverable
from the chokepoint that names it.
