+++ polylogue/storage/sqlite/archive_tiers/source_write.py
@@ -13,7 +13,7 @@ from dataclasses import dataclass, replace
 from datetime import UTC, datetime
 from typing import Literal
 
-from polylogue.archive.revision_authority import RawRevisionAuthority, RawRevisionEnvelope
+from polylogue.archive.revision_authority import RawRevisionAuthority, RawRevisionEnvelope, RawRevisionKind
 from polylogue.core.enums import ArtifactSupportStatus, Origin, Provider, ValidationMode, ValidationStatus
 from polylogue.core.raw_failure_evidence import RAW_FAILURE_EVIDENCE_KINDS
 from polylogue.storage.introspection import table_exists as _table_exists
@@ -556,6 +556,106 @@ def write_history_sidecar(
             (resolved_sidecar_id, origin_value, source_path, payload_json, observed_at_ms, content_hash),
         )
     return resolved_sidecar_id
+
+
+@dataclass(frozen=True, slots=True)
+class ReconstructedRawRow:
+    """Every column of a raw row rebuilt from already-proven evidence."""
+
+    raw_id: str
+    origin: str
+    capture_mode: str | None
+    native_id: str | None
+    source_path: str
+    source_index: int
+    blob_hash: bytes
+    blob_size: int
+    acquired_at_ms: int
+    logical_source_key: str
+    source_revision: str
+    baseline_raw_id: str
+
+
+def insert_reconstructed_raw_row(
+    conn: sqlite3.Connection,
+    row: ReconstructedRawRow,
+    *,
+    schema: str = "main",
+) -> None:
+    """Insert a raw row RECONSTRUCTED from proven evidence -- not an observation.
+
+    This is the one named exemption from :func:`admit_raw_observation`, and it
+    is architectural rather than a not-yet-migrated call site.
+
+    ``admit_raw_observation`` decides *what an observation means*: it takes
+    freshly read bytes plus the accepted head and resolves which revision
+    envelope those bytes earn. Every input to that decision is missing here.
+    A copy-forward repair (``storage/repair.py``'s browser-origin
+    reconstruction) is not observing a source at all -- the source it would
+    observe is gone. It is rewriting evidence the archive already holds and
+    has already adjudicated, under a corrected identity, from a repair plan
+    whose contents were proven before this call. Handing those bytes to the
+    chokepoint would ask it to re-derive a verdict that is an input here, and
+    it would derive a *different* one: with no prior head for the corrected
+    logical key it would resolve BASELINE/``ASSERTED``, discarding the
+    ``BYTE_PROVEN`` authority the repair plan established.
+
+    So the exemption is not "this write skips the typed arms". It is that the
+    arms have already been resolved, durably, elsewhere -- which is why every
+    envelope column is a required field of :class:`ReconstructedRawRow`
+    instead of a nullable this function could leave unset. The invariant the
+    chokepoint exists for (typed resolution, no nullable limbo) holds on this
+    path by construction; what does not apply is the resolution *step*.
+
+    Callers must be repair/migration paths writing under an explicit,
+    receipt-emitting plan. Acquisition routes must use
+    :func:`admit_raw_observation`.
+    """
+    if schema not in {"main", "source"}:
+        raise ValueError(f"unsupported source schema: {schema}")
+    if len(row.blob_hash) != 32:
+        raise ValueError("blob_hash must be a 32-byte SHA-256 digest")
+
+    columns = {str(info[1]) for info in conn.execute(f"PRAGMA {schema}.table_info(raw_sessions)")}
+    names = [
+        "raw_id",
+        "origin",
+        "native_id",
+        "source_path",
+        "source_index",
+        "blob_hash",
+        "blob_size",
+        "acquired_at_ms",
+        "logical_source_key",
+        "revision_kind",
+        "source_revision",
+        "baseline_raw_id",
+        "acquisition_generation",
+        "revision_authority",
+    ]
+    values: list[object] = [
+        row.raw_id,
+        row.origin,
+        row.native_id,
+        row.source_path,
+        row.source_index,
+        row.blob_hash,
+        row.blob_size,
+        row.acquired_at_ms,
+        row.logical_source_key,
+        RawRevisionKind.FULL.value,
+        row.source_revision,
+        row.baseline_raw_id,
+        0,
+        RawRevisionAuthority.BYTE_PROVEN.value,
+    ]
+    if "capture_mode" in columns:
+        names.insert(2, "capture_mode")
+        values.insert(2, row.capture_mode)
+    conn.execute(
+        f"INSERT INTO {schema}.raw_sessions ({', '.join(names)}) VALUES ({', '.join('?' for _ in names)})",
+        values,
+    )
 
 
 def write_source_raw_session(
@@ -1535,6 +1635,8 @@ __all__ = [
     "pending_raw_logical_source_key",
     "upsert_raw_artifact",
     "write_history_sidecar",
+    "ReconstructedRawRow",
+    "insert_reconstructed_raw_row",
     "write_source_raw_session",
     "write_source_raw_session_blob_ref",
 ]
